Update CPU Scheduler
In this version of the main method, I used the Process class that you made to create an array of Process objects to utilize throughout the program. I also made the initialization of the burst times and arrival times of each process object consistent with our assumptions; each process ID is simply the index of a for loop (e.g. Process #1 = 1, Process #2 = 2, etc.). Let me know as soon as possible if I missed anything or if anything needs to be changed. 

Josh Papello