Update README.md
I modified Assumption 7 to state that, if two processes have exactly the same burst time, then the one which is executed first will be the one with the lowest ID number. My FCFS and SJF algorithms still obtain the correct values for the average wait time and the average turnaround time, so changing this assumption should not effect any of the others whatsoever and permit my code to be consistent with every assumption (including this new one!).