Create CPUScheduler2.java
This is the latest version of my CPU Scheduler, complete with working algorithms for FCFS and SJF. Furthermore, I have made my code open to the possibility of us implementing the priority algorithm (higher number meaning higher priority). This is a file separate from the original because I did not want to get rid of the changes you made to the original in case we needed them for later.